Updated package name to support maven groupId syntax
