+++ CSS Parser/pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>self.philbrown.cssparser</groupId>
+  <groupId>io.github.phdbrown.cssparser</groupId>
   <artifactId>jCSS-Parser</artifactId>
   <packaging>jar</packaging>
   <version>1.0-SNAPSHOT</version>
