+++ CSS Parser/src/main/java/io/github/phdbrown/cssparser/App.java
@@ -1,4 +1,4 @@
-package self.philbrown.cssparser;
+package io.github.phdbrown.cssparser;
 
 /**
  * Hello world!
