+++ CSS Parser/src/main/java/io/github/phdbrown/cssparser/KeyFrame.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package self.philbrown.cssparser;
+package io.github.phdbrown.cssparser;
 
 import java.util.ArrayList;
 import java.util.List;
